[EDA] Tag of final questions & Answer Rate
[Feature]
Answer Rate : Correct Answer Rate of for each Tag of final sequence
Tag of final questions : just like the name

[How]
Scatterplots were used to determine correlations between variables.

In addition, firstly, a graph was drawn for the tag answer rate, the number of tags, and the correct answer code, and 0 and 1 for the Answer Code were grouped in different colors in the graph.

Since the results didn't seem to be coming out well, I looked at the correlation between the tag answer rate for the last question and the tag answer rate for the entire data.